% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PakNAcc.R
\name{PakNAcc}
\alias{PakNAcc}
\title{PakNAcc}
\usage{
PakNAcc()
}
\value{
Pivot tables and graphs for Pakistan's Quarterly National Accounts data
}
\description{
Shiny App for National Accounts
}
\examples{
if(interactive()) {
    library(PakNAcc)
    PakNAcc()
  }

}
